Setup the main function of user logic

+++ controller/product-controller.js
@@ -19,70 +19,44 @@ const getAllProduct = async (req, res) => {
 };
 
 const updateProduct = async (req, res) => {
-  let { payload } = req.body;
+  const { productName, productCategory, unitOfMeasure, productPrice } =
+    req.body;
   const { id } = req.params;
   const image = req.file;
-
-  if (typeof payload === "string") payload = JSON.parse(payload);
   let product;
-  if (payload.productCategory) {
-    let category;
-    try {
-      category = await Category.findOne({
-        categoryName: payload.productCategory,
-      });
-    } catch (err) {
-      fs.unlinkSync(image.path);
-      return res
-        .status(402)
-        .json({ message: "No unit of measure match for this product" });
-    }
-    if (category) {
-      payload.productCategory = category;
-    } else {
-      fs.unlinkSync(image.path);
-      return res
-        .status(402)
-        .json({ message: "No unit of measure match for this product" });
-    }
-  }
-  if (payload.unitOfMeasure) {
-    let unitOfMeasure;
-    try {
-      unitOfMeasure = await UnitOfMeasure.findOne({
-        unitOfMeaureName: payload.unitOfMeasure,
-      });
-    } catch (err) {
-      fs.unlinkSync(image.path);
-      return res
-        .status(402)
-        .json({ message: "No unit of measure match for this product" });
-    }
-    if (unitOfMeasure) {
-      payload.unitOfMeasure = unitOfMeasure;
-    } else {
-      fs.unlinkSync(image.path);
-      return res
-        .status(402)
-        .json({ message: "No unit of measure match for this product" });
-    }
-  }
+  let unit = await UnitOfMeasure.findOne({unitOfMeasureName:unitOfMeasure})
+  let cate = await Category.findOne({categoryName:productCategory})
   try {
     if (image) {
       product = await Product.findByIdAndUpdate(
         id,
-        { ...payload, productImage: image.path },
+        {
+          productName,
+          productCategory: cate,
+          unitOfMeasure:unit,
+          productPrice,
+          productImage: image.path,
+        },
         { returnOriginal: true }
       );
       try {
         fs.unlinkSync(product.productImage);
       } catch (err) {}
     } else
-      product = await Product.findByIdAndUpdate(id, payload, {
-        returnOriginal: true,
-      });
+      product = await Product.findByIdAndUpdate(
+        id,
+        {
+          productName,
+          productCategory:cate,
+          unitOfMeasure:unit,
+          productPrice,
+        },
+        {
+          returnOriginal: true,
+        }
+      );
   } catch (err) {
-    fs.unlinkSync(image.path);
+    if (image) fs.unlinkSync(image.path);
     console.log(err);
     return res
       .status(402)
@@ -94,7 +68,6 @@ const updateProduct = async (req, res) => {
   fs.unlinkSync(image.path);
   return res.status(402).json({ message: "No match with this product" });
 };
-
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
   let product;
@@ -114,35 +87,53 @@ const deleteProduct = async (req, res) => {
 const createProduct = async (req, res) => {
   const { productName, productCategory, unitOfMeasure, productPrice } =
     req.body;
+
   const imagePath = req.file;
   let product;
   try {
     product = await Product.findOne({
       productName: productName,
-      unitOfMeasure: unitOfMeasure,
+      unitOfMeasure: await UnitOfMeasure.findOne({
+        unitOfMeaureName: unitOfMeasure,
+      }),
     });
   } catch (err) {
+    fs.unlinkSync(imagePath.path);
     return res
       .status(402)
       .json({ message: "Could not Create new product, please try again" });
   }
   if (!product) {
-    product = new Product({
-      productName: productName,
-      productCategory: await Category.findOne({
-        categoryName: productCategory,
-      }),
-      unitOfMeasure: await UnitOfMeasure.findById(unitOfMeasure),
-      productPrice: productPrice,
-      productImage: imagePath,
-    });
-    product.save();
-    return res.status(201).json({ message: "New product have been created" });
+    try {
+      product = new Product({
+        productName: productName,
+        productCategory: await Category.findOne({
+          categoryName: productCategory,
+        }),
+        unitOfMeasure: await UnitOfMeasure.findOne({
+          unitOfMeasureName: unitOfMeasure,
+        }),
+        productPrice: productPrice,
+        productImage: imagePath.path,
+      });
+      product.save();
+      return res.status(201).json({ id: product.id });
+    } catch (err) {
+      fs.unlinkSync(imagePath.path);
+      return res.status(402).json({ message: "The product is already exists" });
+    }
   }
+  fs.unlinkSync(imagePath.path);
   return res.status(402).json({ message: "The product is already exists" });
+};
+
+const clearAll = async (req, res) => {
+  await Product.deleteMany();
+  return res.status(201).json({ message: "delete all items" });
 };
 
 exports.getAllProduct = getAllProduct;
 exports.deleteProduct = deleteProduct;
 exports.updateProduct = updateProduct;
 exports.createProduct = createProduct;
+exports.clearAll = clearAll;
